heatwave shader and fetch assets handles rugsea crashes

+++ src/Realm/properties/RealmSky/components/MilkyWaySky/materials/milkyway.ts
@@ -44,8 +44,8 @@ const vert = `
 `;
 
 const frag = glsl`
-    precision highp float;
-    precision highp int;
+    precision mediump float;
+    precision mediump int;
     
     uniform float time;
     varying vec2 vUv;
